+++ src/sdk/resources/index.ts
@@ -4,3 +4,4 @@ export * from './invoices';
 export * from './products';
 export * from './customers';
 export * from './shops';
+export * from './blacklist';